Update: utilizing promises for loading scripts
