+++ src/hooks/useScriptLoader.ts
@@ -1,19 +1,5 @@
-import { useCallback, useEffect } from "react";
-
-import {
-  getFromWindowCache,
-  updateCachedScript,
-  CachedScript,
-  addScriptUpdater,
-  removeScriptUpdater,
-} from "../scriptCache";
-
-const getNewScript = (source: string): HTMLScriptElement => {
-  const newScript = document.createElement("script");
-  newScript.async = true;
-  newScript.setAttribute("src", source);
-  return newScript;
-};
+import { useCallback, useEffect, useRef } from "react";
+import { waitForScript } from "../scriptloader-support";
 
 export interface ScriptLoaderConfiguration {
   onSuccess: () => void;
@@ -25,105 +11,35 @@ export interface ScriptLoader {
   (config: ScriptLoaderConfiguration): void;
 }
 
-interface CachedScriptUpdater extends ScriptLoader {}
-
-const useCachedScriptUpdater: CachedScriptUpdater = ({
-  onSuccess,
-  onFailure,
-  source,
-}) => {
-  const updater = useCallback(
-    ({ loading, failed, failureEvent }: CachedScript) => {
-      if (!loading && !failed) {
-        onSuccess();
-      }
-      if (!loading && failed) {
-        onFailure(failureEvent);
-      }
-    },
-    [onSuccess, onFailure]
-  );
-
-  useEffect(() => {
-    addScriptUpdater(source, updater);
-    return () => removeScriptUpdater(source, updater);
-  }, [updater, source]);
-
-  useEffect(() => {
-    // run updater with already cached info
-    updater(getFromWindowCache(source));
-  }, [source, updater]);
-};
-
 const useScriptLoader: ScriptLoader = (config) => {
-  const { source } = config;
-  useCachedScriptUpdater(config);
-
-  const setupListeners = useCallback(
-    (scriptRef: HTMLScriptElement): (() => void) => {
-      const removeListeners = () => {
-        scriptRef.removeEventListener("load", loadEvent);
-        scriptRef.removeEventListener("error", errorEvent);
-      };
-
-      const generateScriptEventListener = (
-        getResultingCachedScript: (ev: Event) => Partial<CachedScript>
-      ) => (ev: Event) => {
-        updateCachedScript(source, getResultingCachedScript(ev));
-        removeListeners();
-      };
-
-      const loadEvent = generateScriptEventListener(() => ({
-        loading: false,
-        failed: false,
-      }));
-
-      const errorEvent = generateScriptEventListener((err: ErrorEvent) => ({
-        loading: false,
-        failed: true,
-        failureEvent: err,
-      }));
-
-      scriptRef.addEventListener("load", loadEvent);
-      scriptRef.addEventListener("error", errorEvent);
-
-      return removeListeners;
+  const {
+    source,
+    onSuccess,
+    onFailure = () => {
+      //noop
     },
-    [source]
+  } = config;
+  const isMounted = useRef(true);
+  useEffect(() => () => (isMounted.current = false));
+  const successFunc = useCallback(() => isMounted.current && onSuccess(), [
+    onSuccess,
+  ]);
+  const errorFunc = useCallback(
+    (err: ErrorEvent) => isMounted.current && onFailure(err),
+    [onFailure]
   );
 
   useEffect(() => {
-    let scriptRef = document.querySelector<HTMLScriptElement>(
-      `script[src="${source}"]`
-    );
-    const scriptExists = Boolean(scriptRef);
-
-    if (scriptExists) {
-      const cachedScriptInfo = getFromWindowCache(source);
-      if (!cachedScriptInfo.scriptCreated) {
-        // if we did not create the script, assume it has loaded
-        updateCachedScript(source, {
-          loading: false,
-          failed: false,
-        });
-        return;
+    const waitForSource = async () => {
+      try {
+        await waitForScript(source);
+        successFunc();
+      } catch (err) {
+        errorFunc(err as ErrorEvent);
       }
-
-      // if we are not loading, do nothing
-      if (!cachedScriptInfo.loading) return;
-
-      // if we are loading and we did create the script, listen
-      return setupListeners(scriptRef);
-    }
-
-    // if we did not create the script, create it
-    scriptRef = getNewScript(source);
-    updateCachedScript(source, { scriptCreated: true });
-    const removeListeners = setupListeners(scriptRef);
-    document.body.appendChild(scriptRef);
-
-    return removeListeners;
-  }, [source, setupListeners]);
+    };
+    void waitForSource();
+  }, [source, successFunc, errorFunc]);
 };
 
 export default useScriptLoader;
